refactor: rename 'Section' folder to 'sections'
