+++ california-dev/src/components/sections/Technical/Technical.jsx
